Refactor Template Elements for Active Records
