+++ modules/template/elements/BaseContentRecordsElement.php
@@ -35,13 +35,13 @@ abstract class BaseContentRecordsElement extends BaseRecordsElement
      */
     protected function getDynamicAttributes(): array
     {
-        return array_merge(parent::getDynamicAttributes(), [
+        return [
             'space' => null,
             'author' => null,
             'topic' => null,
             'filter' => null,
             'limit' => null,
-        ]);
+        ];
     }
 
     /**
@@ -70,61 +70,12 @@ abstract class BaseContentRecordsElement extends BaseRecordsElement
 
     /**
      * @inheritdoc
-     */
-    public function getTypes(): array
-    {
-        return array_merge(parent::getTypes(), [
-            'options' => Yii::t('CustomPagesModule.base', 'Specific criteria'),
-        ]);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function isConfigured(): bool
-    {
-        return parent::isConfigured() || $this->type === 'options';
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function isCacheable(): bool
-    {
-        // Don't cache because the filter `ContentActiveRecord::find()->readable()` is used here
-        return false;
-    }
-
-    /**
-     * @inheritdoc
      * @return ActiveQueryContent
      */
     protected function getQuery(): ActiveQuery
     {
         $query = static::RECORD_CLASS::find()->readable();
 
-        return match ($this->type) {
-            'options' => $this->filterOptions($query),
-            default => $this->filterStatic($query),
-        };
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function filterStatic(ActiveQuery $query): ActiveQuery
-    {
-        return $query->andWhere(['content.object_id' => $this->static]);
-    }
-
-    /**
-     * Filter by options
-     *
-     * @param ActiveQueryContent $query
-     * @return ActiveQuery
-     */
-    protected function filterOptions(ActiveQueryContent $query): ActiveQuery
-    {
         $query->userRelated([
             ActiveQueryContent::USER_RELATED_SCOPE_OWN_PROFILE,
             ActiveQueryContent::USER_RELATED_SCOPE_SPACES,
@@ -159,7 +110,7 @@ abstract class BaseContentRecordsElement extends BaseRecordsElement
             }
         }
 
-        return $query;
+        return $query->limit($this->limit);
     }
 
     /**
@@ -185,13 +136,10 @@ abstract class BaseContentRecordsElement extends BaseRecordsElement
      */
     public function renderEditForm(ActiveForm $form): string
     {
-        return parent::renderEditForm($form) .
-            $this->renderEditRecordsTypeFields([
-                'options' => $form->field($this, 'space')->widget(SpacePickerField::class) .
-                    $form->field($this, 'author')->widget(UserPickerField::class) .
-                    $form->field($this, 'topic')->widget(TopicPicker::class) .
-                    $form->field($this, 'filter')->checkboxList($this->getContentFilterOptions()) .
-                    $form->field($this, 'limit'),
-            ]);
+        return $form->field($this, 'space')->widget(SpacePickerField::class) .
+            $form->field($this, 'author')->widget(UserPickerField::class) .
+            $form->field($this, 'topic')->widget(TopicPicker::class) .
+            $form->field($this, 'filter')->checkboxList($this->getContentFilterOptions()) .
+            $form->field($this, 'limit');
     }
 }
